requirements.txt
pandas~=2.1.3
python-dotenv~=1.0.0
spotipy~=2.23.0

# testing
coverage~=7.3.2
flake8~=6.1.0
pytest~=7.4.2